// types/routes/SettingsRoutes.ts
export type SettingsRoutes =
  | 'Settings'
  | 'Notifications_settings'
  | 'Securite_et_privee'
  | 'Contact_et_FAQ'
  | 'Aide'
  | 'Centre_de_securite'
  | 'Nous_contactez'
  | 'FAQ'
  | 'Recherche'
  | 'Emplacement'
  | 'Mode_invisible'
  | 'Mode_voyage'
  | 'Mettre_en_pause'
  | 'Mode_de_connexion'
  | 'Changer_localisation'
  | 'Parametre_de_confidentialites'
  | 'Autorisations_necessaires'
  | 'Bloquer_contacts'
  | 'Supprimer_mon_compte'
  | 'Compte_non_trouve';